Sprite animation system for eventual pathfinder character

+++ AI/AStar/source/VectorMaths.cpp
@@ -21,5 +21,5 @@ sf::Vector2f normalize(sf::Vector2f& input)
 float angleBetweenVectors(sf::Vector2f leftHandVector, sf::Vector2f rightHandVector)
 {
     sf::Vector2f angleVector = leftHandVector - rightHandVector;
-    return std::atan2f(angleVector.y, angleVector.x) * 180 / M_PI;
+    return std::atan2f(angleVector.y, angleVector.x) * 180 / (float)M_PI;
 }
