+++ AI/Agent/source/copyMousePosition.cpp
@@ -5,5 +5,5 @@
 void copyMousePosition::Update(Agent *pAgent)
 {
     sf::Vector2i mousePos = sf::Mouse::getPosition(Renderer::getInstance()->getWindow());
-    pAgent->setPosition(sf::Vector2f(mousePos.x, mousePos.y));
+    pAgent->setPosition(sf::Vector2f((float)mousePos.x, (float)mousePos.y));
 }
